+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Next version
+
+* Add TV season translations method by @laurentblue in https://github.com/UweTrottmann/tmdb-java/pull/108
+
 ## 2.12.0 - 2025-08-08
 
 * Note: TMDB has added additional primary translation languages: `ku-TR`, `so-SO` and `uz-UZ`.